% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab_methods.R
\name{apa_table.h_tab}
\alias{apa_table.h_tab}
\title{apa_table.h_tab}
\usage{
apa_table.h_tab(
  x,
  caption = NULL,
  note = NULL,
  stub_indents = NULL,
  added_stub_head = NULL,
  col_spanners = NULL,
  midrules = NULL,
  placement = "tbp",
  landscape = FALSE,
  font_size = NULL,
  escape = FALSE,
  span_text_columns = TRUE,
  format.args = NULL,
  ...
)
}
\arguments{
\item{x}{passed to papaja::apa_table}

\item{caption}{passed to papaja::apa_table}

\item{note}{passed to papaja::apa_table}

\item{stub_indents}{passed to papaja::apa_table}

\item{added_stub_head}{passed to papaja::apa_table}

\item{col_spanners}{passed to papaja::apa_table}

\item{midrules}{passed to papaja::apa_table}

\item{placement}{passed to papaja::apa_table}

\item{landscape}{passed to papaja::apa_table}

\item{font_size}{passed to papaja::apa_table}

\item{escape}{passed to papaja::apa_table}

\item{span_text_columns}{passed to papaja::apa_table}

\item{format.args}{passed to papaja::apa_table}

\item{...}{passed to papaja::apa_table}
}
\description{
htab method for apa_table
}
