% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab_part.R
\name{h_table_part.lm}
\alias{h_table_part.lm}
\title{h_tab}
\usage{
h_table_part.lm(
  m1,
  m0 = NULL,
  model_name = "untitled model",
  ci_method = "NULL",
  transf = NULL,
  transf_name = NULL
)
}
\arguments{
\item{m1}{A model to describe}

\item{m0}{a nested, less complicated model}

\item{model_name}{The display name of m1}
}
\description{
Present results including a model comparison
}
\examples{
m0  <- lm(mpg ~ hp, data = mtcars)
m1 <- lm(mpg ~ hp + wt, data = mtcars)
h_tab(m0, m1)
}
