% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab.R
\name{h_table_part}
\alias{h_table_part}
\title{h_tab}
\usage{
h_table_part(
  m1,
  m0 = NULL,
  model_name,
  ci_method = NULL,
  transf = NULL,
  transf_name = NULL
)
}
\arguments{
\item{m1}{A model to describe}

\item{m0}{a nested, less complicated model}

\item{model_name}{The display name of m1}
}
\description{
h_tab
}
\examples{
m0  <- lm(mpg ~ hp, data = mtcars)
m1 <- lm(mpg ~ hp + wt, data = mtcars)
h_tab(m1, m0, model_name = "Horsepower and Weight")
}
