% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab.R
\name{tab_model.glm}
\alias{tab_model.glm}
\title{tab_model.glm}
\usage{
tab_model.glm(model, ci_method = NULL, transf = NULL, transf_name = NULL)
}
\arguments{
\item{model}{glm model}

\item{ci_method}{confidence interval method send to parameters::parameters}

\item{transf}{optional transformation function}

\item{transf_name}{optional transformation name}
}
\description{
tab_model method for general linear models
}
