% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab.R
\name{tab_model.lmerMod}
\alias{tab_model.lmerMod}
\title{tab_model.lmerMod}
\usage{
tab_model.lmerMod(model, ci_method = NULL, transf = NULL, transf_name = NULL)
}
\arguments{
\item{model}{lmerMod model}

\item{ci_method}{confidence interval method send to parameters::parameters}

\item{transf}{optional transformation function}

\item{transf_name}{optional transformation name}
}
\description{
tab_model method for lmerMod models
}
