% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab.R
\name{h_tab}
\alias{h_tab}
\title{h_tab}
\usage{
h_tab(..., ci_method = NULL, transf = NULL, transf_name = NULL)
}
\arguments{
\item{...}{a list of models, otherwise named models}

\item{ci_method}{what method should be used to derive confidence intervals?}

\item{transf}{function to transform coef and 95\% CI}

\item{transf_name}{what should the transformed column be called?}
}
\description{
h_tab
}
