% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_tab.R
\name{tab}
\alias{tab}
\title{h_tab}
\usage{
tab(model, ci_method = NULL, transf = NULL, transf_name = NULL, replace = NULL)
}
\arguments{
\item{model}{model to tabulate}

\item{ci_method}{passed to parameters}

\item{transf}{function to transform coefficients (not SE)}

\item{transf_name}{name for transformed coefficient}

\item{replace}{named vector, names replaced with contents}

\item{m1}{A model to describe}

\item{m0}{a nested, less complicated model}

\item{model_name}{The display name of m1}
}
\description{
h_tab
}
\examples{
m0  <- lm(mpg ~ hp, data = mtcars)
m1 <- lm(mpg ~ hp + wt, data = mtcars)
h_tab(m0, m1)
tab

Assemble model info into data.frame
}
